attempt to fix code duplicate
